PtypeIterator moved to a separate file
